fix: use leader units in integration tests, not unit 0 (#70)
* fix: use leader units in integration tests, not unit 0

Previously, when the integration tests interacted with a unit of a deployed application they used the hardcoded unit 0.  This typically works, but sometimes tests would flake if an application did not have a unit 0 (maybe because 0 was killed and 1 was created).

This change adds `get_leader_unit_number` to first determined the unit number of the leader unit, then uses that unit anywhere we previously used unit 0.  This should be more resilient while functionally equivalent (as before, we used unit 0 because that's typically the leader/only unit)